아반떼 2010~2023

+++ src/main/java/com/example/helloCar/domain/hellocar/domain/car/Car.java
@@ -20,10 +20,10 @@ public class Car {
     private String carName;
     //차 이름
 
-    private Long minPrice;
+    private int minPrice;
     //최소 가격
 
-    private Long maxPrice;
+    private int maxPrice;
     //최대 가격
 
     private String vehicle;
